+++ Sprint_1/version1/data/users.csv
@@ -1,4 +1,4 @@
-username,password,ID,Full Name,Role
+username,password,id,name,role
 sasas,asasasa,instructor,asasa,sasasa
 username,password,instructor,id,name
 ee3e3username,e3e3e3epassword,instructor,e3e3eid,fullname
